requirements.txt
Django==5.2.10
django-colorfield==0.14.0
djangorestframework==3.16.1
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
